Update framework-multiple example
