+++ examples/framework-multiple/.editorconfig
@@ -3,7 +3,7 @@ root = true
 [*]
 charset = utf-8
 end_of_line = lf
-indent_style = tab
+indent_style = space
 indent_size = 2
 insert_final_newline = true
 trim_trailing_whitespace = true
