+++ examples/framework-multiple/.gitignore
@@ -1,12 +1,23 @@
-.nx/cache
-.nx/workspace-data
-bin/
-dist/
-node_modules/
-obj/
+# Logs
+logs
+*.log
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
+pnpm-debug.log*
+lerna-debug.log*
 
-*.tmp
-*.tsbuildinfo
-lerna-debug.log
-TODO.md
-vite.config.ts.*
+node_modules
+dist
+dist-ssr
+bin/
+obj/
+*.local
+
+# Editor directories and files
+.idea
+.DS_Store
+*.suo
+*.ntvs*
+*.njsproj
+*.sw?
