+++ examples/framework-multiple/.nvmrc
@@ -1,1 +1,1 @@
-v22.11.0
+v22.x
