+++ examples/framework-multiple/.vscode/extensions.json
@@ -1,10 +1,10 @@
 {
-	"recommendations": [
-		"aaron-bond.better-comments",
-		"biomejs.biome",
-		"EditorConfig.EditorConfig",
-		"ms-dotnettools.csdevkit",
-		"svelte.svelte-vscode",
-		"Vue.volar"
-	]
+  "recommendations": [
+    "biomejs.biome",
+    "DotJoshJohnson.xml",
+    "EditorConfig.EditorConfig",
+    "ms-dotnettools.csdevkit",
+    "svelte.svelte-vscode",
+    "Vue.volar"
+  ]
 }
