+++ examples/framework-multiple/FrameworkMultiple.sln
