+++ examples/framework-multiple/README.md
@@ -1,51 +1,65 @@
-# Multiple frameworks example
+# Multiple UI frameworks example
 
-Phoria Islands using multiple UI frameworks (React, Svelte and Vue) in the same project.
+This example demonstrates how to use multiple UI frameworks in a single project.
 
-## Getting started
-
-The example can be run in development or preview (production build) mode.
-
-### Run in development mode
+You can use this example as a template for your own project by running:
 
 ```shell
-# Change into the WebApp directory
-cd ./WebApp
-
-# Install dependencies
-corepack install
-pnpm install
-
-# Create dotnet dev cert
-dotnet dev-certs https --trust
-
-# Run Phoria dev server
-pnpm run dev
-
-# Run Phoria web app (in different terminal window/tab)
-dotnet run
-
-# The web app should be running at https://localhost:5246/
+npx giget@latest gh:cmeeg/phoria-examples/examples/framework-multiple <target_dir>
 ```
 
-### Run in preview mode
+> [!IMPORTANT]
+> You will need to replace:
+> * `<target_dir>` with the name of the local directory you want to clone the example project to
+
+## Usage
+
+Once cloned you will need to install the dependencies:
 
 ```shell
-# Change into the WebApp directory
-cd ./WebApp
-
-# Install dependencies
-corepack install
+corepack enable pnpm
 pnpm install
+```
 
-# Create dotnet dev cert
+Then you can run the project in dev mode:
+
+```shell
+# Add dev certs
 dotnet dev-certs https --trust
 
-# Create production build
-pnpm run build
+# Start the Phoria Server
+pnpm dev
 
-# Run Phoria web app
-pnpm run preview
-
-# The web app should be running at http://localhost:5245/
+# Start the Phoria Web App
+# You will need to run this in a separate terminal instance/tab to the Phoria Server
+dotnet run --project WebApp/WebApp.csproj --launch-profile Development
 ```
+
+Or build the project for production:
+
+```shell
+# Build the project
+pnpm build
+
+# Preview the production build
+pnpm preview
+```
+
+Or run the production build in a Docker container:
+
+```shell
+# Build the container image
+docker build -f ./WebApp/Dockerfile -t phoriaapp:latest .
+
+# Run the container image (and browse on http://localhost:3001)
+docker run --name phoriaapp -d -p 3001:8080 phoriaapp:latest
+
+# Stop the container image
+docker stop phoriaapp
+
+# Remove the container image
+docker rm phoriaapp
+```
+
+> [!NOTE]
+> You will need to have [Docker Desktop](https://docs.docker.com/desktop/) installed before running the above commands.
