+++ examples/framework-multiple/WebApp/Pages/Error.cshtml.cs
@@ -8,20 +8,20 @@ namespace WebApp.Pages;
 [IgnoreAntiforgeryToken]
 public class ErrorModel : PageModel
 {
-	public string? RequestId { get; set; }
+    public string? RequestId { get; set; }
 
-	public bool ShowRequestId => !string.IsNullOrEmpty(RequestId);
+    public bool ShowRequestId => !string.IsNullOrEmpty(RequestId);
 
-	private readonly ILogger<ErrorModel> _logger;
+    private readonly ILogger<ErrorModel> _logger;
 
-	public ErrorModel(ILogger<ErrorModel> logger)
-	{
-		_logger = logger;
-	}
+    public ErrorModel(ILogger<ErrorModel> logger)
+    {
+        _logger = logger;
+    }
 
-	public void OnGet()
-	{
-		RequestId = Activity.Current?.Id ?? HttpContext.TraceIdentifier;
-	}
+    public void OnGet()
+    {
+        RequestId = Activity.Current?.Id ?? HttpContext.TraceIdentifier;
+    }
 }
