+++ examples/framework-multiple/WebApp/Pages/Index.cshtml
@@ -1,32 +1,32 @@
 ﻿@page
 @model IndexModel
 @{
-	ViewData["Title"] = "Home page";
+    ViewData["Title"] = "Home page";
 }
 
 <div>
-	<a href="https://github.com/CMeeg/phoria" target="_blank" rel="noreferrer">
-		<img src="~/img/phoria.svg" class="logo" alt="Phoria logo" asp-append-version="true" />
-	</a>
+    <a href="https://github.com/CMeeg/phoria" target="_blank" rel="noreferrer">
+        <img src="~/img/phoria.svg" class="logo" alt="Phoria logo" />
+    </a>
 </div>
 <h1>Phoria Islands</h1>
 @if (Model.ShowReactComponent)
 {
-	<div class="card">
-		<phoria-island component="ReactCounter" client="Client.Load" props="new { StartAt = 5 }"></phoria-island>
-	</div>
+    <div class="card">
+        <phoria-island component="ReactCounter" client="Client.Load" props="new { StartAt = 5 }"></phoria-island>
+    </div>
 }
 @if (Model.ShowVueComponent)
 {
-	<div class="card">
-		<phoria-island component="VueCounter" client="Client.Idle()" props="new { StartAt = 10 }"></phoria-island>
-	</div>
+    <div class="card">
+        <phoria-island component="VueCounter" client="Client.Idle()" props="new { StartAt = 10 }"></phoria-island>
+    </div>
 }
 @if (Model.ShowSvelteComponent)
 {
-	<div class="card">
-		<phoria-island component="SvelteCounter" client="@(Client.Visible("50px"))"
-			props="new { StartAt = 15 }"></phoria-island>
-	</div>
+    <div class="card">
+        <phoria-island component="SvelteCounter" client="@(Client.Visible("50px"))"
+            props="new { StartAt = 15 }"></phoria-island>
+    </div>
 }
 <p class="read-the-docs">Click on the logos to learn more</p>
