+++ examples/framework-multiple/WebApp/Pages/Index.cshtml.cs
@@ -5,26 +5,26 @@ namespace WebApp.Pages;
 
 public class IndexModel : PageModel
 {
-	private readonly ILogger<IndexModel> logger;
+    private readonly ILogger<IndexModel> logger;
 
-	public bool ShowReactComponent { get; set; } = true;
-	public bool ShowVueComponent { get; set; } = true;
-	public bool ShowSvelteComponent { get; set; } = true;
+    public bool ShowReactComponent { get; set; } = true;
+    public bool ShowVueComponent { get; set; } = true;
+    public bool ShowSvelteComponent { get; set; } = true;
 
-	public IndexModel(ILogger<IndexModel> logger)
-	{
-		this.logger = logger;
-	}
+    public IndexModel(ILogger<IndexModel> logger)
+    {
+        this.logger = logger;
+    }
 
-	public void OnGet([FromQuery] string[]? framework)
-	{
-		if (framework == null || framework.Length == 0)
-		{
-			return;
-		}
+    public void OnGet([FromQuery] string[]? framework)
+    {
+        if (framework == null || framework.Length == 0)
+        {
+            return;
+        }
 
-		ShowReactComponent = framework.Contains("react", StringComparer.OrdinalIgnoreCase);
-		ShowVueComponent = framework.Contains("vue", StringComparer.OrdinalIgnoreCase);
-		ShowSvelteComponent = framework.Contains("svelte", StringComparer.OrdinalIgnoreCase);
-	}
+        ShowReactComponent = framework.Contains("react", StringComparer.OrdinalIgnoreCase);
+        ShowVueComponent = framework.Contains("vue", StringComparer.OrdinalIgnoreCase);
+        ShowSvelteComponent = framework.Contains("svelte", StringComparer.OrdinalIgnoreCase);
+    }
 }
