+++ examples/framework-multiple/WebApp/Pages/Shared/_Layout.cshtml
@@ -1,22 +1,25 @@
 ﻿<!DOCTYPE html>
 <html lang="en">
+
 <head>
-	<meta charset="utf-8" />
-	<meta name="viewport" content="width=device-width, initial-scale=1.0" />
-	<title>@ViewData["Title"] - WebApp</title>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>@ViewData["Title"] - WebApp</title>
 
-	<phoria-island-styles />
-	<phoria-island-preload />
+    <phoria-island-styles />
+    <phoria-island-preload />
 
-	@await RenderSectionAsync("Styles", required: false)
+    @await RenderSectionAsync("Styles", required: false)
 </head>
+
 <body>
-	<main role="main" id="root">
-		@RenderBody()
-	</main>
+    <main role="main" id="root">
+        @RenderBody()
+    </main>
 
-	<phoria-island-scripts />
+    <phoria-island-scripts />
 
-	@await RenderSectionAsync("Scripts", required: false)
+    @await RenderSectionAsync("Scripts", required: false)
 </body>
+
 </html>
