+++ examples/framework-multiple/WebApp/Pages/_ViewStart.cshtml
@@ -1,3 +1,3 @@
 ﻿@{
-	Layout = "_Layout";
+    Layout = "_Layout";
 }
