+++ examples/framework-multiple/WebApp/Properties/launchSettings.json
@@ -1,23 +1,23 @@
 {
-	"$schema": "https://json.schemastore.org/launchsettings.json",
-	"profiles": {
-		"Development": {
-			"commandName": "Project",
-			"dotnetRunMessages": true,
-			"launchBrowser": true,
-			"applicationUrl": "https://localhost:5246;http://localhost:5245",
-			"environmentVariables": {
-				"DOTNET_ENVIRONMENT": "Development"
-			}
-		},
-		"Preview": {
-			"commandName": "Project",
-			"dotnetRunMessages": true,
-			"launchBrowser": true,
-			"applicationUrl": "http://localhost:5245",
-			"environmentVariables": {
-				"DOTNET_ENVIRONMENT": "Preview"
-			}
-		}
-	}
+  "$schema": "https://json.schemastore.org/launchsettings.json",
+  "profiles": {
+    "Development": {
+      "commandName": "Project",
+      "dotnetRunMessages": true,
+      "launchBrowser": true,
+      "applicationUrl": "https://localhost:5246;http://localhost:5245",
+      "environmentVariables": {
+        "DOTNET_ENVIRONMENT": "Development"
+      }
+    },
+    "Preview": {
+      "commandName": "Project",
+      "dotnetRunMessages": true,
+      "launchBrowser": true,
+      "applicationUrl": "http://localhost:5247",
+      "environmentVariables": {
+        "DOTNET_ENVIRONMENT": "Preview"
+      }
+    }
+  }
 }
