+++ examples/framework-multiple/WebApp/appsettings.Development.json
@@ -1,8 +1,14 @@
 {
-	"detailedErrors": true,
-	"phoria": {
-		"server": {
-			"https": true
-		}
-	}
+  "DetailedErrors": true,
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft.AspNetCore": "Warning"
+    }
+  },
+  "phoria": {
+    "server": {
+      "https": true
+    }
+  }
 }
