+++ examples/framework-multiple/WebApp/appsettings.Preview.json
@@ -1,3 +1,3 @@
 {
-	"detailedErrors": true
+  "DetailedErrors": true
 }
