+++ examples/framework-multiple/WebApp/appsettings.Production.json
@@ -1,10 +1,10 @@
 {
-	"phoria": {
-		"server": {
-			"process": {
-				"command": "node",
-				"arguments": ["ui/dist/server/server.js"]
-			}
-		}
-	}
+  "phoria": {
+    "server": {
+      "process": {
+        "command": "node",
+        "arguments": ["WebApp/ui/dist/server/server.js"]
+      }
+    }
+  }
 }
