+++ examples/framework-multiple/WebApp/appsettings.json
@@ -1,14 +1,15 @@
 {
-	"Logging": {
-		"LogLevel": {
-			"Default": "Information",
-			"Microsoft.AspNetCore": "Warning",
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft.AspNetCore": "Warning",
 			"System.Net.Http.HttpClient.Phoria.Vite.DevServer.DevHttpClient": "Warning"
-		}
-	},
-	"allowedHosts": "*",
-	"phoria": {
-		"entry": "src/entry-client.ts",
-		"ssrEntry": "src/entry-server.ts"
-	}
+    }
+  },
+  "AllowedHosts": "*",
+  "phoria": {
+    "root": "WebApp/ui",
+    "entry": "src/entry-client.ts",
+    "ssrEntry": "src/entry-server.ts"
+  }
 }
