--- examples/framework-multiple/WebApp/package.json
@@ -1,47 +1,0 @@
-{
-	"name": "webapp",
-	"private": true,
-	"version": "0.0.0",
-	"type": "module",
-	"scripts": {
-		"build": "run-p build:* -c",
-		"build:app": "vite build --app",
-		"build:dotnet": "dotnet build --configuration Release",
-		"build:server": "vite build --config vite.server.config.ts",
-		"check": "tsc",
-		"dev": "tsx ./ui/src/server.ts",
-		"lint": "biome check",
-		"preview": "run-p preview:* -c",
-		"preview:dotnet": "cross-env DOTNET_ENVIRONMENT=preview dotnet run ./WebApp/bin/Release/net9.0/WebApp.dll --launch-profile Preview",
-		"preview:phoria": "cross-env NODE_ENV=production DOTNET_ENVIRONMENT=Preview node ./ui/dist/server/server.js"
-	},
-	"dependencies": {
-		"@phoria/phoria": "^0.4.1",
-		"@phoria/phoria-react": "^0.4.1",
-		"@phoria/phoria-svelte": "^0.3.1",
-		"@phoria/phoria-vue": "^0.3.1",
-		"h3": "^1.13.0",
-		"listhen": "^1.9.0",
-		"react": "^19.0.0",
-		"react-dom": "^19.0.0",
-		"svelte": "^5.17.3",
-		"vue": "^3.5.13"
-	},
-	"devDependencies": {
-		"@phoria/vite-plugin-dotnet-dev-certs": "0.2.0",
-		"@sveltejs/vite-plugin-svelte": "^5.0.0",
-		"@types/node": "^22.10.5",
-		"@types/react": "^19.0.4",
-		"@types/react-dom": "^19.0.2",
-		"@vitejs/plugin-react": "^4.3.4",
-		"@vitejs/plugin-vue": "^5.0.3",
-		"cross-env": "^7.0.3",
-		"npm-run-all": "^4.1.5",
-		"postcss-preset-env": "^10.1.3",
-		"tsx": "^4.19.2",
-		"typescript": "~5.7.3",
-		"vite": "^6.0.7",
-		"vite-tsconfig-paths": "^5.1.4"
-	},
-	"packageManager": "pnpm@9.15.0+sha512.76e2379760a4328ec4415815bcd6628dee727af3779aaa4c914e3944156c4299921a89f976381ee107d41f12cfa4b66681ca9c718f0668fa0831ed4c6d8ba56c"
-}
