--- examples/framework-multiple/WebApp/svelte.config.js
@@ -1,7 +1,0 @@
-import { vitePreprocess } from "@sveltejs/vite-plugin-svelte"
-
-export default {
-	// Consult https://svelte.dev/docs#compile-time-svelte-preprocess
-	// for more information about preprocessors
-	preprocess: vitePreprocess()
-}
