--- examples/framework-multiple/WebApp/tsconfig.json
@@ -1,29 +1,0 @@
-{
-	"compilerOptions": {
-		"allowImportingTsExtensions": true,
-		"baseUrl": ".",
-		"esModuleInterop": true,
-		"isolatedModules": true,
-		"jsx": "react-jsx",
-		"lib": ["ES2022", "DOM", "DOM.Iterable"],
-		"module": "ESNext",
-		"moduleDetection": "force",
-		"moduleResolution": "Bundler",
-		"noEmit": true,
-		"noFallthroughCasesInSwitch": true,
-		"noUncheckedSideEffectImports": true,
-		"noUnusedLocals": true,
-		"noUnusedParameters": true,
-		"paths": {
-			"~/*": ["ui/src/*"]
-		},
-		"resolveJsonModule": true,
-		"skipLibCheck": true,
-		"sourceMap": true,
-		"strict": true,
-		"target": "ES2022",
-		"useDefineForClassFields": true,
-		"verbatimModuleSyntax": true
-	},
-	"include": ["ui/src/**/*.ts", "ui/src/**/*.d.ts", "ui/src/**/*.tsx", "ui/src/**/*.vue", "ui/src/**/*.svelte"]
-}
