--- examples/framework-multiple/WebApp/tsconfig.server.json
@@ -1,21 +1,0 @@
-{
-	"compilerOptions": {
-		"baseUrl": ".",
-		"esModuleInterop": true,
-		"isolatedModules": true,
-		"lib": ["ES2022"],
-		"module": "NodeNext",
-		"moduleResolution": "NodeNext",
-		"noEmit": true,
-		"noFallthroughCasesInSwitch": true,
-		"noUncheckedSideEffectImports": true,
-		"noUnusedLocals": true,
-		"noUnusedParameters": true,
-		"resolveJsonModule": true,
-		"skipLibCheck": true,
-		"strict": true,
-		"sourceMap": true,
-		"target": "ES2022"
-	},
-	"include": ["ui/src/server.ts", "vite.config.ts", "vite.server.config.ts"]
-}
