+++ examples/framework-multiple/WebApp/ui/src/components/Counter/Counter.module.css
@@ -1,25 +1,25 @@
 .counter {
-	button {
-		display: block;
-		margin: auto;
-	}
+  button {
+    display: block;
+    margin: auto;
+  }
 
-	:global(.logo.react):hover {
-		filter: drop-shadow(0 0 2em #61dafbaa);
-	}
+  :global(.logo.react):hover {
+    filter: drop-shadow(0 0 2em #61dafbaa);
+  }
 
-	@media (prefers-reduced-motion: no-preference) {
-		:global(.logo.react) {
-			animation: logo-spin infinite 20s linear;
-		}
-	}
+  @media (prefers-reduced-motion: no-preference) {
+    :global(.logo.react) {
+      animation: logo-spin infinite 20s linear;
+    }
+  }
 }
 
 @keyframes logo-spin {
-	from {
-		transform: rotate(0deg);
-	}
-	to {
-		transform: rotate(360deg);
-	}
+  from {
+    transform: rotate(0deg);
+  }
+  to {
+    transform: rotate(360deg);
+  }
 }
