+++ examples/framework-multiple/WebApp/ui/src/components/Counter/Counter.svelte
@@ -9,23 +9,23 @@ let count = $state(startAt)
 </script>
 
 <div class="svelte-counter">
-	<a href="https://svelte.dev" target="_blank" rel="noreferrer">
-		<img src="{base}/svelte.svg" class="logo svelte" alt="Svelte Logo" />
-	</a>
-	<button type="button" onclick={() => count++}>count is {count}</button>
-	<p>
-		Edit
-		<code>ui/src/components/Counter/Counter.svelte</code> to test HMR
-	</p>
+  <a href="https://svelte.dev" target="_blank" rel="noreferrer">
+    <img src="{base}/svelte.svg" class="logo svelte" alt="Svelte Logo" />
+  </a>
+  <button type="button" onclick={() => count++}>count is {count}</button>
+  <p>
+    Edit
+    <code>ui/src/components/Counter/Counter.svelte</code> to test HMR
+  </p>
 </div>
 
 <style>
 .logo.svelte:hover {
-	filter: drop-shadow(0 0 2em #ff3e00aa);
+  filter: drop-shadow(0 0 2em #ff3e00aa);
 }
 
 button {
-	display: block;
-	margin: auto;
+  display: block;
+  margin: auto;
 }
 </style>
