+++ examples/framework-multiple/WebApp/ui/src/components/Counter/Counter.vue
@@ -7,25 +7,25 @@ const count = ref(props.startAt ?? 0)
 </script>
 
 <template>
-	<div class="vue-counter">
-		<a href="https://vuejs.org/" target="_blank" rel="noreferrer">
-			<img src="/vue.svg" class="logo vue" alt="Vue logo" />
-		</a>
-		<button type="button" @click="count++">count is {{ count }}</button>
-		<p>
-			Edit
-			<code>ui/src/components/Counter/Counter.vue</code> to test HMR
-		</p>
-	</div>
+  <div class="vue-counter">
+    <a href="https://vuejs.org/" target="_blank" rel="noreferrer">
+      <img src="/vue.svg" class="logo vue" alt="Vue logo" />
+    </a>
+    <button type="button" @click="count++">count is {{ count }}</button>
+    <p>
+      Edit
+      <code>ui/src/components/Counter/Counter.vue</code> to test HMR
+    </p>
+  </div>
 </template>
 
 <style scoped>
 .logo.vue:hover {
-	filter: drop-shadow(0 0 2em #42b883aa);
+  filter: drop-shadow(0 0 2em #42b883aa);
 }
 
 button {
-	display: block;
-	margin: auto;
+  display: block;
+  margin: auto;
 }
 </style>
