+++ examples/framework-multiple/WebApp/ui/src/components/register.ts
@@ -1,19 +1,19 @@
 import { registerComponents } from "@phoria/phoria"
 
 registerComponents({
-	ReactCounter: {
-		loader: {
-			module: () => import("./Counter/Counter.tsx"),
-			component: (module) => module.Counter
-		},
-		framework: "react"
-	},
-	VueCounter: {
-		loader: () => import("./Counter/Counter.vue"),
-		framework: "vue"
-	},
-	SvelteCounter: {
-		loader: () => import("./Counter/Counter.svelte"),
-		framework: "svelte"
-	}
+  ReactCounter: {
+    loader: {
+      module: () => import("./Counter/Counter.tsx"),
+      component: (module) => module.Counter
+    },
+    framework: "react"
+  },
+  VueCounter: {
+    loader: () => import("./Counter/Counter.vue"),
+    framework: "vue"
+  },
+  SvelteCounter: {
+    loader: () => import("./Counter/Counter.svelte"),
+    framework: "svelte"
+  }
 })
