+++ examples/framework-multiple/WebApp/ui/src/entry-server.ts
@@ -5,7 +5,7 @@ import "./components/register"
 import type { PhoriaIsland } from "@phoria/phoria/server"
 
 async function renderPhoriaIsland(island: PhoriaIsland) {
-	return await island.render()
+  return await island.render()
 }
 
 export { renderPhoriaIsland }
