--- examples/framework-multiple/WebApp/vite.server.config.ts
@@ -1,23 +1,0 @@
-import { parsePhoriaAppSettings } from "@phoria/phoria/server"
-import { type UserConfig, defineConfig } from "vite"
-
-export default defineConfig(async () => {
-	const dotnetEnv = process.env.DOTNET_ENVIRONMENT ?? process.env.ASPNETCORE_ENVIRONMENT ?? "Development"
-	const appsettings = await parsePhoriaAppSettings({ environment: dotnetEnv })
-
-	// https://vite.dev/config/
-	return {
-		root: appsettings.root,
-		base: appsettings.base,
-		build: {
-			ssr: true,
-			target: "es2022",
-			copyPublicDir: false,
-			emptyOutDir: true,
-			outDir: `${appsettings.build.outDir}/server`,
-			rollupOptions: {
-				input: `${appsettings.root}/src/server.ts`
-			}
-		}
-	} satisfies UserConfig
-})
