+++ examples/framework-multiple/postcss.config.cjs
@@ -1,11 +1,11 @@
 const postcssPresetEnv = require("postcss-preset-env")
 
 const config = () => ({
-	plugins: [
-		postcssPresetEnv({
-			stage: 0
-		})
-	]
+  plugins: [
+    postcssPresetEnv({
+      stage: 0
+    })
+  ]
 })
 
 module.exports = config
