+++ examples/framework-multiple/vite.config.ts
@@ -6,18 +6,14 @@ import { dotnetDevCerts } from "@phoria/vite-plugin-dotnet-dev-certs"
 import { defineConfig } from "vite"
 import tsconfigPaths from "vite-tsconfig-paths"
 
-// https://vite.dev/config/
 export default defineConfig({
-	publicDir: "public",
-	plugins: [
-		tsconfigPaths({
-			// The tsconfig is in the root of the project, not the "Vite root"
-			root: "../"
-		}),
-		dotnetDevCerts(),
-		phoria(),
-		phoriaReact(),
-		phoriaSvelte(),
-		phoriaVue()
-	]
+  publicDir: "public",
+  plugins: [
+    tsconfigPaths({ root: "../../" }),
+    dotnetDevCerts(),
+    phoria({ cwd: "WebApp" }),
+    phoriaReact(),
+    phoriaSvelte(),
+    phoriaVue()
+  ]
 })
